Added Internet Check on flutternetworkservicehandler class.

+++ lib/src/http_network_service.dart
@@ -1,23 +1,42 @@
 import 'dart:convert';
 
+import 'package:data_connection_checker/data_connection_checker.dart';
 import 'package:http/http.dart' as http;
 
 import 'http_exception.dart';
 
 class HttpNetworkService {
+  static _getInternetForMarkAttendance({var seconds}) async {
+    bool result = await DataConnectionChecker().hasConnection.timeout(
+          Duration(seconds: seconds),
+        );
+    if (result == true) {
+      return true;
+    } else {
+      return false;
+    }
+  }
+
   //it sends a get request using http package with exception handling
-  static getRequest({var url, var headers}) async {
+  static getRequest(
+      {var url, var headers, var networkCheckDurationInSeconds}) async {
     try {
-      final response = await http.get(url, headers: headers);
-      if (response.statusCode == 200 || response.statusCode <= 299) {
-        final decodedResponse = json.decode(response.body);
-        return decodedResponse;
-      } else if (response.statusCode == 300 || response.statusCode <= 399) {
-        throw HttpException('Redirection error');
-      } else if (response.statusCode == 400 || response.statusCode <= 499) {
-        throw HttpException('Bad Request Format');
-      } else if (response.statusCode == 500 || response.statusCode <= 599) {
-        throw HttpException('Internal Server Error');
+      bool isInternetAvailable = await _getInternetForMarkAttendance(
+          seconds: networkCheckDurationInSeconds);
+      if (isInternetAvailable) {
+        final response = await http.get(url, headers: headers);
+        if (response.statusCode == 200 || response.statusCode <= 299) {
+          final decodedResponse = json.decode(response.body);
+          return decodedResponse;
+        } else if (response.statusCode == 300 || response.statusCode <= 399) {
+          throw HttpException('Redirection error');
+        } else if (response.statusCode == 400 || response.statusCode <= 499) {
+          throw HttpException('Bad Request Format');
+        } else if (response.statusCode == 500 || response.statusCode <= 599) {
+          throw HttpException('Internal Server Error');
+        }
+      } else {
+        throw HttpException('No Internet Found');
       }
     } catch (e) {
       throw e;
@@ -25,18 +44,28 @@ class HttpNetworkService {
   }
 
   //it sends a post request using http package with exception handling
-  static postRequest({var url, var headers, var body}) async {
+  static postRequest(
+      {var url,
+      var headers,
+      var body,
+      var networkCheckDurationInSeconds}) async {
     try {
-      final response = await http.post(url, headers: headers, body: body);
-      if (response.statusCode == 200 || response.statusCode <= 299) {
-        final decodedResponse = json.decode(response.body);
-        return decodedResponse;
-      } else if (response.statusCode == 300 || response.statusCode <= 399) {
-        throw HttpException('Redirection error');
-      } else if (response.statusCode == 400 || response.statusCode <= 499) {
-        throw HttpException('Bad Request Format');
-      } else if (response.statusCode == 500 || response.statusCode <= 599) {
-        throw HttpException('Internal Server Error');
+      bool isInternetAvailable = await _getInternetForMarkAttendance(
+          seconds: networkCheckDurationInSeconds);
+      if (isInternetAvailable) {
+        final response = await http.post(url, headers: headers, body: body);
+        if (response.statusCode == 200 || response.statusCode <= 299) {
+          final decodedResponse = json.decode(response.body);
+          return decodedResponse;
+        } else if (response.statusCode == 300 || response.statusCode <= 399) {
+          throw HttpException('Redirection error');
+        } else if (response.statusCode == 400 || response.statusCode <= 499) {
+          throw HttpException('Bad Request Format');
+        } else if (response.statusCode == 500 || response.statusCode <= 599) {
+          throw HttpException('Internal Server Error');
+        }
+      } else {
+        throw HttpException('No Internet Found');
       }
     } catch (e) {
       throw e;
@@ -44,18 +73,28 @@ class HttpNetworkService {
   }
 
   //it sends a put request using http package with exception handling
-  static putRequest({var url, var headers, var body}) async {
+  static putRequest(
+      {var url,
+      var headers,
+      var body,
+      var networkCheckDurationInSeconds}) async {
     try {
-      final response = await http.put(url, headers: headers, body: body);
-      if (response.statusCode == 200 || response.statusCode <= 299) {
-        final decodedResponse = json.decode(response.body);
-        return decodedResponse;
-      } else if (response.statusCode == 300 || response.statusCode <= 399) {
-        throw HttpException('Redirection error');
-      } else if (response.statusCode == 400 || response.statusCode <= 499) {
-        throw HttpException('Bad Request Format');
-      } else if (response.statusCode == 500 || response.statusCode <= 599) {
-        throw HttpException('Internal Server Error');
+      bool isInternetAvailable = await _getInternetForMarkAttendance(
+          seconds: networkCheckDurationInSeconds);
+      if (isInternetAvailable) {
+        final response = await http.put(url, headers: headers, body: body);
+        if (response.statusCode == 200 || response.statusCode <= 299) {
+          final decodedResponse = json.decode(response.body);
+          return decodedResponse;
+        } else if (response.statusCode == 300 || response.statusCode <= 399) {
+          throw HttpException('Redirection error');
+        } else if (response.statusCode == 400 || response.statusCode <= 499) {
+          throw HttpException('Bad Request Format');
+        } else if (response.statusCode == 500 || response.statusCode <= 599) {
+          throw HttpException('Internal Server Error');
+        }
+      } else {
+        throw HttpException('No Internet Found');
       }
     } catch (e) {
       throw e;
@@ -63,18 +102,25 @@ class HttpNetworkService {
   }
 
   //it sends a delete request using http package with exception handling
-  static deleteRequest({var url, var headers}) async {
+  static deleteRequest(
+      {var url, var headers, var networkCheckDurationInSeconds}) async {
     try {
-      final response = await http.delete(url, headers: headers);
-      if (response.statusCode == 200 || response.statusCode <= 299) {
-        final decodedResponse = json.decode(response.body);
-        return decodedResponse;
-      } else if (response.statusCode == 300 || response.statusCode <= 399) {
-        throw HttpException('Redirection error');
-      } else if (response.statusCode == 400 || response.statusCode <= 499) {
-        throw HttpException('Bad Request Format');
-      } else if (response.statusCode == 500 || response.statusCode <= 599) {
-        throw HttpException('Internal Server Error');
+      bool isInternetAvailable = await _getInternetForMarkAttendance(
+          seconds: networkCheckDurationInSeconds);
+      if (isInternetAvailable) {
+        final response = await http.delete(url, headers: headers);
+        if (response.statusCode == 200 || response.statusCode <= 299) {
+          final decodedResponse = json.decode(response.body);
+          return decodedResponse;
+        } else if (response.statusCode == 300 || response.statusCode <= 399) {
+          throw HttpException('Redirection error');
+        } else if (response.statusCode == 400 || response.statusCode <= 499) {
+          throw HttpException('Bad Request Format');
+        } else if (response.statusCode == 500 || response.statusCode <= 599) {
+          throw HttpException('Internal Server Error');
+        }
+      } else {
+        throw HttpException('No Internet Found');
       }
     } catch (e) {
       throw e;
@@ -82,18 +128,28 @@ class HttpNetworkService {
   }
 
   //it sends a patch request using http package with exception handling
-  static patchRequest({var url, var headers, var body}) async {
+  static patchRequest(
+      {var url,
+      var headers,
+      var body,
+      var networkCheckDurationInSeconds}) async {
     try {
-      final response = await http.patch(url, headers: headers, body: body);
-      if (response.statusCode == 200 || response.statusCode <= 299) {
-        final decodedResponse = json.decode(response.body);
-        return decodedResponse;
-      } else if (response.statusCode == 300 || response.statusCode <= 399) {
-        throw HttpException('Redirection error');
-      } else if (response.statusCode == 400 || response.statusCode <= 499) {
-        throw HttpException('Bad Request Format');
-      } else if (response.statusCode == 500 || response.statusCode <= 599) {
-        throw HttpException('Internal Server Error');
+      bool isInternetAvailable = await _getInternetForMarkAttendance(
+          seconds: networkCheckDurationInSeconds);
+      if (isInternetAvailable) {
+        final response = await http.patch(url, headers: headers, body: body);
+        if (response.statusCode == 200 || response.statusCode <= 299) {
+          final decodedResponse = json.decode(response.body);
+          return decodedResponse;
+        } else if (response.statusCode == 300 || response.statusCode <= 399) {
+          throw HttpException('Redirection error');
+        } else if (response.statusCode == 400 || response.statusCode <= 499) {
+          throw HttpException('Bad Request Format');
+        } else if (response.statusCode == 500 || response.statusCode <= 599) {
+          throw HttpException('Internal Server Error');
+        }
+      } else {
+        throw HttpException('No Internet Found');
       }
     } catch (e) {
       throw e;
